+++ src/components/about/about.js
@@ -18,6 +18,7 @@ const COMPONENT_NAME = 'about';
  * @param {string} [settings.productName] Additional product name information to display.
  * @param {boolean} [settings.useDefaultCopyright=true] Add the Legal Approved Infor Copyright Text.
  * @param {string} [settings.version] Semantic Version Number for example (4.0.0).
+ * @param {array|object} [settings.showCopyButton=true] Show copy to clipboard button
  * @param {array|object} [settings.attributes=null] Add extra attributes like id's to the toast element. For example `attributes: { name: 'id', value: 'my-unique-id' }`
 */
 const ABOUT_DEFAULTS = {
@@ -28,6 +29,7 @@ const ABOUT_DEFAULTS = {
   productName: undefined,
   useDefaultCopyright: true,
   version: undefined,
+  showCopyButton: true
 };
 
 function About(element, settings) {
@@ -148,11 +150,40 @@ About.prototype = {
       attributes: this.settings.attributes
     });
 
+    if (this.settings.showCopyButton) {
+      $(`<button type="button" class="btn-icon" id="copy-to-clipboard" title="${Locale.translate('CopyToClipboard')}">
+          <span>${Locale.translate('CopyToClipboard')}</span>
+          <svg role="presentation" aria-hidden="true" focusable="false" class="icon">
+            <use href="#icon-copy"></use>
+          </svg>
+        </button>`).prependTo(this.modal.find('.modal-body-wrapper')).on('click', () => {
+        this.copyToClipBoard();
+      }).tooltip();
+    }
+
     // Link the About API to the Modal API
     const modalAPI = this.modal.data('modal');
     modalAPI.aboutAPI = this;
 
     return this;
+  },
+
+  /**
+   * Copy inner text to clipboard
+   */
+  copyToClipBoard() {
+    const container = document.createElement('div');
+    container.innerHTML = this.modal.find('.modal-body').html();
+    container.style.position = 'fixed';
+    container.style.pointerEvents = 'none';
+    container.style.opacity = 0;
+    document.body.appendChild(container);
+    window.getSelection().removeAllRanges();
+    const range = document.createRange();
+    range.selectNode(container);
+    window.getSelection().addRange(range);
+    document.execCommand('copy');
+    document.body.removeChild(container);
   },
 
   /**