+++ src/utils/utils.js
@@ -413,7 +413,6 @@ $.copyToClipboard = function (text) { // eslint-disable-line
     try {
       return document.execCommand('copy'); // Security exception may be thrown by some browsers.
     } catch (ex) {
-      // console.warn('Copy to clipboard failed.', ex);
       return false;
     } finally {
       document.body.removeChild(textarea);